Create lonelyInteger
Problem Statement

There are N integers in an array A. All but one integer occur in pairs. Your task is to find the number that occurs only once.

Input Format

The first line of the input contains an integer N, indicating the number of integers. The next line contains N space-separated integers that form the array A.

input
9
4 9 95 93 57 4 57 93 9
output
95

(9) above means how many numbers will be in
and 95 is the only one that appear only once

source :
https://www.hackerrank.com/challenges/lonely-integer